requirements.txt
asgiref==3.11.0
Django==6.0.1
django-auditlog==3.4.1
django-htmx>=1.21.0
django-tailwind==4.4.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5

# Phase 3: External Integrations
cryptography>=44.0.0
PyGithub>=2.5.0
docker>=7.1.0
django-tasks[database]>=0.4.0
django-formtools>=2.5.0

# Phase 4: Blueprints
semver>=3.0.0
GitPython>=3.1.0
PyYAML>=6.0.0
